+++ src/is_same_iso_week/index.js
@@ -22,7 +22,7 @@ var isSameWeek = require('../is_same_week')
  * //=> true
  */
 function isSameISOWeek (dirtyDateLeft, dirtyDateRight) {
-  return isSameWeek(dirtyDateLeft, dirtyDateRight, {weekStartsAt: 1})
+  return isSameWeek(dirtyDateLeft, dirtyDateRight, {weekStartsOn: 1})
 }
 
 module.exports = isSameISOWeek