+++ src/is_same_week/index.js
@@ -10,7 +10,7 @@ var startOfWeek = require('../start_of_week')
  * @param {Date|String|Number} dateLeft - the first date to check
  * @param {Date|String|Number} dateRight - the second date to check
  * @param {Object} [options] - the object with options
- * @param {Number} [options.weekStartsAt=0] - the index of the first day of a week (0 - sunday)
+ * @param {Number} [options.weekStartsOn=0] - the index of the first day of a week (0 - sunday)
  * @returns {Boolean} the dates are in the same week
  *
  * @example
@@ -27,7 +27,7 @@ var startOfWeek = require('../start_of_week')
  * var result = isSameWeek(
  *   new Date(2014, 7, 31),
  *   new Date(2014, 8, 4),
- *   {weekStartsAt: 1}
+ *   {weekStartsOn: 1}
  * )
  * //=> false
  */