+++ src/last_day_of_week/test.js
@@ -14,7 +14,7 @@ describe('lastDayOfWeek', function () {
 
   it('allows to specify which day is the first day of the week', function () {
     var date = new Date(2014, 8 /* Sep */, 2, 11, 55, 0)
-    var result = lastDayOfWeek(date, {weekStartsAt: 1})
+    var result = lastDayOfWeek(date, {weekStartsOn: 1})
     assert.deepEqual(result,
       new Date(2014, 8 /* Sep */, 7)
     )
@@ -46,7 +46,7 @@ describe('lastDayOfWeek', function () {
     context('when the given day is before the start of a week', function () {
       it('it returns the last day of a week', function () {
         var date = new Date(2014, 9 /* Oct */, 6)
-        var result = lastDayOfWeek(date, {weekStartsAt: 3})
+        var result = lastDayOfWeek(date, {weekStartsOn: 3})
         assert.deepEqual(result, new Date(2014, 9 /* Oct */, 7))
       })
     })
@@ -54,7 +54,7 @@ describe('lastDayOfWeek', function () {
     context('when the given day is the start of a week', function () {
       it('it returns the last day of a week', function () {
         var date = new Date(2014, 9 /* Oct */, 8)
-        var result = lastDayOfWeek(date, {weekStartsAt: 3})
+        var result = lastDayOfWeek(date, {weekStartsOn: 3})
         assert.deepEqual(result, new Date(2014, 9 /* Oct */, 14))
       })
     })
@@ -62,14 +62,14 @@ describe('lastDayOfWeek', function () {
     context('when the given day is after the start of a week', function () {
       it('it returns the last day of a week', function () {
         var date = new Date(2014, 9 /* Oct */, 10)
-        var result = lastDayOfWeek(date, {weekStartsAt: 3})
+        var result = lastDayOfWeek(date, {weekStartsOn: 3})
         assert.deepEqual(result, new Date(2014, 9 /* Oct */, 14))
       })
     })
 
     it('handles the week at the end of a year', function () {
       var date = new Date(2014, 11 /* Dec */, 29)
-      var result = lastDayOfWeek(date, {weekStartsAt: 5})
+      var result = lastDayOfWeek(date, {weekStartsOn: 5})
       assert.deepEqual(result, new Date(2015, 0 /* Jan */, 1))
     })
   })